+++ tiger-common/src/main/java/de/gematik/test/tiger/common/config/TigerConfigurationSource.java
@@ -36,7 +36,6 @@ public class TigerConfigurationSource {
             .collect(Collectors.toList());
         appliedTemplates.forEach(key -> finalValues.remove(key));
 
-
         return finalValues;
     }
 }