+++ tiger-common/src/main/java/de/gematik/test/tiger/common/config/TigerGlobalConfiguration.java
@@ -16,6 +16,10 @@ public class TigerGlobalConfiguration {
     }
 
     public synchronized static void initialize() {
+        if (initialized) {
+            return;
+        }
+
         initialized = true;
         globalConfigurationLoader.initialize();
     }